# perf: faster package-lock tree creation