--- a/lib/set-immediate-promise.ts
+++ b/lib/set-immediate-promise.ts
@@ -1,0 +1,3 @@
+export function setImmediatePromise() {
+  return new Promise((resolve) => setImmediate(resolve));
+}