--- a/package.json
+++ b/package.json
@@ -30,9 +30,11 @@
   "homepage": "https://github.com/snyk/nodejs-lockfile-parser#readme",
   "dependencies": {
     "@yarnpkg/lockfile": "^1.0.2",
+    "graphlib": "^2.1.5",
     "lodash": "4.17.10",
     "source-map-support": "^0.5.7",
-    "tslib": "^1.9.3"
+    "tslib": "^1.9.3",
+    "uuid": "^3.3.2"
   },
   "devDependencies": {
     "@types/lodash": "^4.14.116",