--- a/test/lib/fixtures/cyclic-dep-self-reference/cyclic-dep-self-reference-2/package.json
+++ b/test/lib/fixtures/cyclic-dep-self-reference/cyclic-dep-self-reference-2/package.json
@@ -1,0 +1,14 @@
+{
+  "name": "cyclic-dep-self-reference-2",
+  "version": "1.0.0",
+  "description": "",
+  "main": "index.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "author": "",
+  "license": "ISC",
+  "dependencies": {
+    "cyclic-dep-self-reference": "file:../../cyclic-dep-self-reference"
+  }
+}