--- a/test/lib/fixtures/cyclic-dep-self-reference/package.json
+++ b/test/lib/fixtures/cyclic-dep-self-reference/package.json
@@ -1,0 +1,14 @@
+{
+  "name": "cyclic-dep-self-reference",
+  "version": "1.0.0",
+  "main": "index.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "author": "",
+  "license": "ISC",
+  "description": "",
+  "dependencies": {
+    "cyclic-dep-self-reference-2": "file:./cyclic-dep-self-reference-2"
+  }
+}