🚧 **DO NOT MERGE** 🚧

- [x] Tests written and linted [ℹ︎]()
- [ ] Documentation written [ℹ︎]()
- [x] Commit history is tidy [ℹ︎]()

### What this does

Big `package-lock.json` parsing was slow. This PR introduces better parsing algorithm, which speeds up parsing a lot. 

_Instructions on how to run this locally, background context, what to review, questions…_
Proper testing is still needed to be sure that functionality is the same. Tests are passing*, but they can't cover everything. However, it's pretty hard to verify what's right for the output with ~million lines.

\* I am aware about one failing, it's a topic for the discussion about the meaning of `depType` property.

*Before:*
```
$ time parse-nodejs-lockfile > /dev/null
parse-nodejs-lockfile > /dev/null  30.25s user 1.24s system 123% cpu 25.522 total
```
*After:*
```
$ time node ../lockfile-parser/bin > /dev/null
node ../lockfile-parser/bin > /dev/null  0.39s user 0.06s system 103% cpu 0.436 total
```

*30.25s* vs  *0.39s*