@CLAassistant 2018-10-03
[![CLA assistant check](https://cla-assistant.io/pull/badge/not_signed)](https://cla-assistant.io/snyk/nodejs-lockfile-parser?pullRequest=20) <br/>Thank you for your submission, we really appreciate it. Like many open source projects, we ask that you sign our [Contributor License Agreement](https://cla-assistant.io/snyk/nodejs-lockfile-parser?pullRequest=20) before we can accept your contribution.<br/><sub>You have signed the CLA already but the status is still pending? Let us [recheck](https://cla-assistant.io/check/snyk/nodejs-lockfile-parser?pullRequest=20) it.</sub>