@miiila 2018-11-02
@michael-go I did my best to add a test with self-reference cycle. This was so far the best I was able to create, because I'm not aware about any existing package, which would transitively require itself (`trucolor` doesn't do it anymore, it transitively requires older version of itself 😂); since it has an impact on how packages are hoisted, it's not easy to manually adjuste `package-lock.json` to simulate this situation. 

So I created that kind of package by myself and algorithm doesn't break - it identifies cycle one level lower than it really is, but that's the limitation I'm happy to live with.