@snyksec 2018-11-05
:tada: This PR is included in version 1.7.0 :tada:

The release is available on:
- [npm package (@latest dist-tag)](https://www.npmjs.com/package/snyk-nodejs-lockfile-parser)
- [GitHub release](https://github.com/snyk/nodejs-lockfile-parser/releases/tag/v1.7.0)

Your **[semantic-release](https://github.com/semantic-release/semantic-release)** bot :package::rocket: